requirements.txt
python-frontmatter>=0.5.0
python-dateutil>=2.8.1
pytz==2020.1
git+https://github.com/lukasschwab/jsonfeed.git
